requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
hjson==3.0.1
idna==2.8
python-vlc==3.0.7110
requests==2.22.0
soupsieve==1.9.5
urllib3==1.25.7
